Add a way of defining default service versions automatically
A new ConfigMap is created with the default image and ETR versions
defined in it. This file is really easy to modify in a github workflow
which is why it is created.
The ConfigMap will be deployed together with the ETOS controller which
I like because it will be easy to verify the defaults from within
Kubernetes.
The values in the ConfigMap are then used to replace the default tags
in the Cluster CRD so that they are used when the values are not set
in spec.
Added a new github action which reads the latest docker images from
our service repositories as well as the latest release from the ETR
repository. The action then adds these versions into the ConfigMap,
updating the defaults.